executor: move FillVirtualColumnValue to the table package
