+++ executor/point_get.go
@@ -347,8 +347,8 @@ func (e *PointGetExecutor) Next(ctx context.Context, req *chunk.Chunk) error {
 		return err
 	}
 
-	err = FillVirtualColumnValue(e.virtualColumnRetFieldTypes, e.virtualColumnIndex,
-		e.schema, e.columns, e.ctx, req)
+	err = table.FillVirtualColumnValue(e.virtualColumnRetFieldTypes, e.virtualColumnIndex,
+		e.schema.Columns, e.columns, e.ctx, req)
 	if err != nil {
 		return err
 	}
